requirements.txt
langchain-core
langchain-community
pydantic
python-dotenv
pydantic-settings
dashscope
faiss-cpu
uvicorn
watchfiles
python-jose[cryptography]
passlib[bcrypt]
python-multipart
slowapi
structlog
prometheus-client
sentence-transformers
sqlalchemy
aiomysql
pymysql
pytest
pytest-asyncio
httpx
pytest-cov
oss2>=2.18.0
apscheduler>=3.10.0
jieba>=0.42.1
rank-bm25>=0.2.2

# 生产稳定化依赖
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-exporter-jaeger>=1.20.0
opentelemetry-exporter-prometheus>=0.41b0
circuitbreaker>=2.0.0
tenacity>=8.2.0
pybreaker>=1.0.2

# 阶段2依赖
chromadb>=0.4.0
pymilvus>=2.3.0